requirements.txt
asgiref==3.9.1
crispy-tailwind==1.0.3
Django==6.0.1
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-crispy-forms==2.5
django-filter==25.1
django-js-asset==3.1.2
django-modeltranslation==0.19.19
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
pywin32==311
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.45.1
